+++ pkg/ebsnvme/ebsnvme_test.go
@@ -15,18 +15,6 @@ func TestGetVolumeID(t *testing.T) {
 	assert.Equal(t, "vol-67890", i.getVolumeID())
 }
 
-func TestGetDevicePath(t *testing.T) {
-	i := &nvmeIdentifyController{}
-	copy(i.vs.bdev[:], "foo                             ")
-	assert.Equal(t, "/dev/foo", i.getDevicePath())
-
-	copy(i.vs.bdev[:], "/dev/foo                        ")
-	assert.Equal(t, "/dev/foo", i.getDevicePath())
-
-	copy(i.vs.bdev[:], "foobar                          ")
-	assert.Equal(t, "/dev/foobar", i.getDevicePath())
-}
-
 func TestGetDeviceName(t *testing.T) {
 	i := &nvmeIdentifyController{}
 	copy(i.vs.bdev[:], "foo                             ")