+++ CHANGELOG.md
@@ -11,6 +11,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 - Default installation location from `/usr/local/bin` to `/usr/local/sbin`
 
+### Added
+
+- New output fields: `device-namespace` & `device-partition`
+
 ## [v0.2.0] - 2023-10-13
 
 ### Changed